chore(lib/utils): relocate debounce function
