+++ src/components/Admin/Settings.jsx
@@ -2,7 +2,7 @@ import { useTranslation } from "react-i18next";
 import "@/i18n/i18n";
 import ThemeSwitcher from "@/components/Admin/ThemeSwitcher";
 import InfoTooltip from "@/components/ui/tooltip";
-import { debounce } from "../../utils";
+import { debounce } from "@/lib/utils";
 
 function FinalRoundTitleChanger(props) {
   const { i18n, t } = useTranslation();
